Move stream tests to st2stream package.
